task_service/requirements.txt
# FastAPI server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database Driver (MongoDB)
motor==3.3.2
pymongo==4.6.1

# Config
python-dotenv==1.0.1

# Security (Reading JWTs)
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.9.2
email-validator==2.2.0

# Inter-Service Communication
httpx==0.27.0

# NEW: needed for UploadFile / Form
python-multipart==0.0.9
